requirements.txt
altgraph==0.17.4
Appium-Python-Client==1.0.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Deprecated==1.2.14
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
img2pdf==0.5.1
importlib_metadata==7.1.0
lxml==5.2.2
outcome==1.3.0.post0
packaging==24.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pefile==2023.2.7
pikepdf==8.15.1
pillow==10.3.0
pycparser==2.22
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pypdf==4.2.0
pypdfium2==4.30.0
PySocks==1.7.1
pywin32-ctypes==0.2.2
selenium==3.141.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==1.26.2
wrapt==1.16.0
wsproto==1.2.0
zipp==3.18.2
